Romove unnecessary 'units_on_order' from permitted_attributes
'units_on_order' is an atrribute which holds amount of units of given
product that was already ordered. When creating new product its value
is obviously 0, and when modyfing product data manually we do not want
to change this value not to make the data inconsistent. This is why
the crete/update form does not have corresponding field. This makes
'units_on_order' attribute in 'permitted_attributes' unneeded.